moved models into one file and added persistence interface
