--- src/incident_reports/incident_reports_server/model/Severity.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-class Severity(Enum):
-    LOW = "Low"
-    Medium = "Medium"
-    High = "High"
-    Critical = "Critical"
-    
-
