--- src/incident_reports/incident_reports_server/model/Status.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-class Status(Enum):
-    OPEN = "Open"
-    CLOSED = "Closed"
-    RESOLVED = "Resolved"
-    IN_PROGRESS = "In Progress"
